+++ src/config/middleware.js
@@ -1,5 +1,6 @@
 module.exports = {
   DEFAULT: 'app',
+  GLOBAL: 'global',
   APP: 'app',
   ERROR: 'error',
   NOT_FOUND: 'notfound'